Edited the digitize file to now work with the geant output and the generator output to (1) take in geant output and do reconstructions, (2) take in the generator output to make electron reconstructions for the matching electron and (3) read the true values for momentum from the generator output.
I made a new class called Recon_Event do track all this.

Also for the electrons, which just have a TYPE, TRUEMOM, RECONMOM, the other values (pos / barNo / time / energy) were all set to -2147483648. I do not want 0's in there because it interferes with the other distributions for the neutorn. This is a bit redundant since you can insist you only plot for type="neutron", which will take away these 0's, but just in case this is forgotten, it is now clear